move rest api stuff into xos directory
