+++ planetstack/core/xoslib/methods/plus.py
@@ -2,7 +2,7 @@ from rest_framework import generics
 from rest_framework import serializers
 from rest_framework.response import Response
 from rest_framework import status
-from xosapibase import XOSRetrieveUpdateDestroyAPIView, XOSListCreateAPIView
+from xos.apibase import XOSRetrieveUpdateDestroyAPIView, XOSListCreateAPIView
 
 """ PlusSerializerMixin
